fix(mod): Apply MyPage Title
